+ user controller: get patients & remove patient methods

+++ SmartDoctor.User/Core/UserRepository.cs
@@ -1,9 +1,11 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using SmartDoctor.Data.ContextModels;
 using SmartDoctor.Data.Enums;
+using SmartDoctor.Data.JsonModels;
 using SmartDoctor.Helper;
-using SmartDoctor.User.Models;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace SmartDoctor.User.Core
@@ -16,24 +18,31 @@ namespace SmartDoctor.User.Core
         {
             _context = context;
         }
-        public async Task<bool> Authorize(string phoneNumber, string password)
+        public async Task<Users> Authorize(string phoneNumber, string password)
         {
             var person = await _context.Users.FirstOrDefaultAsync(p => p.PhoneNumber.Equals(phoneNumber));
             if (person == null)
                 throw new Exception("Person not found");
-            return SecurePasswordHasher.ValidatePassword(password, person.Password);
-        }      
+            if (SecurePasswordHasher.ValidatePassword(password, person.Password))
+                return person;
+            else
+                throw new Exception("wrong password");
+        }
 
         public async Task RemoveUser(long userId)
         {
             var user = await _context.Users.FirstOrDefaultAsync(x => x.UserId == userId);
             if (user == null)
                 throw new Exception("User not found");
+            var patient = await _context.Patients.FirstOrDefaultAsync(x => x.UserId == userId);
+            if (patient == null)
+                throw new Exception("Patient not found");
             _context.Users.Remove(user);
+            _context.Patients.Remove(patient);
             await _context.SaveChangesAsync();
         }
 
-        public async Task Registration(Users user)
+        public async Task Registration(PatientModel user)
         {
             var userInDb = await _context.Users.FirstOrDefaultAsync(x => x.PhoneNumber == user.PhoneNumber);
             if (userInDb != null)
@@ -43,22 +52,32 @@ namespace SmartDoctor.User.Core
                 CreatedDate = DateTime.UtcNow,
                 Password = SecurePasswordHasher.HashPassword(user.Password),
                 PhoneNumber = user.PhoneNumber,
-                Role = user.Role ?? (byte)RoleTypes.Patient
+                Role = (byte)RoleTypes.Patient
             };
             _context.Users.Add(newUser);
             await _context.SaveChangesAsync();
-        }
-
-        public async Task UpdateUserData(Users user)
-        {
-            var userCtx = await _context.Users.FirstOrDefaultAsync(x => x.UserId == user.UserId);
-            if (userCtx == null)
-                throw new Exception("Person not found");
-            userCtx.Password = user.Password;
-            userCtx.PhoneNumber = user.PhoneNumber;
-            userCtx.Role = user.Role;
+            var patient = new Patients
+            {
+                DateBirth = user.DateBirth,
+                Fio = user.Fio,
+                Gender = bool.Parse(user.Gender),
+                WorkPlace = user.WorkPlace,
+                UserId = newUser.UserId
+            };
+            _context.Patients.Add(patient);
             await _context.SaveChangesAsync();
         }
+
+        //public async Task UpdateUserData(PatientModel model)
+        //{
+        //    var userCtx = await _context.Patients.FirstOrDefaultAsync(x => x. == model.);
+        //    if (userCtx == null)
+        //        throw new Exception("Person not found");
+        //    userCtx.Password = user.Password;
+        //    userCtx.PhoneNumber = user.PhoneNumber;
+        //    userCtx.Role = user.Role;
+        //    await _context.SaveChangesAsync();
+        //}
 
         public async Task<Users> GetUserById(long id)
         {
@@ -80,5 +99,19 @@ namespace SmartDoctor.User.Core
                 throw new Exception("User not found");
             return (RoleTypes)user.Role;
         }
+
+        public async Task<IEnumerable<PatientModel>> GetPatients() =>
+            await (from p in _context.Patients
+                   join u in _context.Users on p.UserId equals u.UserId
+                   select new PatientModel
+                   {
+                       UserId = u.UserId.ToString(),
+                       DateBirth = p.DateBirth,
+                       Fio = p.Fio,
+                       Gender = p.Gender ? "male" : "female",
+                       Password = u.Password,
+                       PhoneNumber = u.PhoneNumber,
+                       WorkPlace = p.WorkPlace
+                   }).ToListAsync();
     }
 }
